updated README file

+++ Genomes/hg38/get_hg38.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-wget http://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/hg38.chromFa.tar.gz
-tar xvfz hg38.chromFa.tar.gz 
-find chroms -type f | grep -v "_" | perl -e 'while(<>){chomp;system("mv $_ .");}'
-rm -r chroms/ hg38.chromFa.tar.gz
-cat *.fa > hg38.fa
-rm chr*.fa
+if [ ! -f hg38.fa ]; then
+    wget http://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/hg38.chromFa.tar.gz
+    tar xvfz hg38.chromFa.tar.gz 
+    find chroms -type f | grep -v "_" | perl -e 'while(<>){chomp;system("mv $_ .");}'
+    rm -r chroms/ hg38.chromFa.tar.gz
+    cat *.fa > hg38.fa
+    faidx hg38.fa -i chromsizes > hg38.chrom.sizes
+fi
